Add text injection to TokenFileReader, that is tokenized and returned before the rest of the input Integrate with the new interface of TokenLine

+++ src/z80asm/z80asm2/macros.h
@@ -11,7 +11,7 @@
 #include <string>
 #include <vector>
 
-extern int g_unique_id_counter; // to create unique local names
+extern size_t g_unique_id_counter; // to create unique local names
 
 //-----------------------------------------------------------------------------
 // Macro
